Almost done repeating payments
Still have to fix delete behaviour. Decided against cronjobs, went for
on login crunching that user's data, might be faster for an individual
user if we do cronjobs, however we can easily change it.

+++ xmlhttp/delete.php
@@ -4,11 +4,19 @@
 	opendb();
 	
 	if($loggedin==1){
-		$id=mysql_real_escape_string($_GET['id']);
-		$query="DELETE FROM payments WHERE PaymentID='$id' AND UserID='$user'";
-		mysql_query($query) or die(mysql_error());
+		$id=sanitise('id');
+		$query="SELECT * FROM payments WHERE PaymentID='$id' AND UserID='$user'";
+		$result=mysql_query($query) or die(mysql_error());
+		$row=mysql_fetch_assoc($result);
+		if($row['Repeated']==0){
+			$query="DELETE FROM payments WHERE PaymentID='$id' AND UserID='$user'";
+			mysql_query($query) or die(mysql_error());
+		}else{
+			$query="UPDATE payments SET Deleted='1' WHERE PaymentID='$id' AND UserID='$user'";
+			mysql_query($query) or die(mysql_error());
+		}
 		
-		$query="SELECT * FROM payments WHERE UserID='$user'";
+		$query="SELECT * FROM payments WHERE UserID='$user' AND Deleted='0'";
 		$result=mysql_query($query) or die(mysql_error());
 		$total=0;
